refactor: DTO 성격의 객체들 패키지 이동
