+++ src/main/kotlin/com/example/hhpluslectureapply/domain/lecture/LectureApplyHistoryRepository.kt
@@ -1,5 +1,6 @@
 package com.example.hhpluslectureapply.domain.lecture
 
+import com.example.hhpluslectureapply.domain.lecture.dto.LectureApplyHistoryDto
 import com.example.hhpluslectureapply.infrastructure.lecture.entity.LectureApplyHistory
 
 interface LectureApplyHistoryRepository {
