+++ src/main/kotlin/com/example/hhpluslectureapply/infrastructure/lecture/LectureApplyHistoryRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.example.hhpluslectureapply.infrastructure.lecture
 
-import com.example.hhpluslectureapply.domain.lecture.LectureApplyHistoryDto
 import com.example.hhpluslectureapply.domain.lecture.LectureApplyHistoryRepository
+import com.example.hhpluslectureapply.domain.lecture.dto.LectureApplyHistoryDto
 import com.example.hhpluslectureapply.infrastructure.lecture.entity.LectureApplyHistory
 import org.springframework.stereotype.Repository
