+++ src/main/kotlin/com/example/hhpluslectureapply/infrastructure/lecture/LectureRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.example.hhpluslectureapply.infrastructure.lecture
 
-import com.example.hhpluslectureapply.domain.lecture.LectureDto
 import com.example.hhpluslectureapply.domain.lecture.LectureRepository
+import com.example.hhpluslectureapply.domain.lecture.dto.LectureDto
 import com.example.hhpluslectureapply.infrastructure.lecture.entity.Lecture
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Repository
