+++ src/main/kotlin/com/example/hhpluslectureapply/usecase/lecture/LectureFacade.kt
@@ -1,8 +1,10 @@
 package com.example.hhpluslectureapply.usecase.lecture
 
-import com.example.hhpluslectureapply.domain.lecture.LectureApplyHistoryDto
 import com.example.hhpluslectureapply.domain.lecture.LectureApplyHistoryService
 import com.example.hhpluslectureapply.domain.lecture.LectureService
+import com.example.hhpluslectureapply.domain.lecture.dto.LectureApplyHistoryDto
+import com.example.hhpluslectureapply.usecase.lecture.dto.LectureApplyInfo
+import com.example.hhpluslectureapply.usecase.lecture.dto.LectureInfo
 import org.springframework.stereotype.Component
 import java.time.LocalDateTime
